First pass at adding validation to non-scalar meta.
This is very much "works for me" only but shows the flow of validation. Currently there's no sanitization or escaping but those can be added. There is a little forgiveness in some areas (ie some things default to string rather than throw an error) but objects are strictly held to their schema properties keys.